Move shell script to Node.js and update README.md
The conversion of a shell script to a Node.js script is to make sure
that everything can be run on all operating systems.  The .gitignore was
updated to reflect the new commands needed to be run.  More links and
some grammar/formatting changes was also added to the README.