Iss119 Add slider to change fade speed
Attach a slider on the DMX page to control the speed of the fade. The
previous default fade was very slow. The new fade default is faster with
the ability to increase it further with the slider. The slider changes how
much the brightness level changes every time interval so that increasing
the slider value will make the fade happen faster and vice versa.